runn/requirements.txt
Django==2.2
django-ckeditor==5.9.0
django-ckeditor-filebrowser-filer==0.3.0
django-filer==1.6.0
django-js-asset==1.2.2
django-mptt==0.11.0
django-polymorphic==2.0.3
djangorestframework==3.11.0
easy-thumbnails==2.7
Pillow==7.0.0
pkg-resources==0.0.0
psycopg2==2.8.4
pytz==2019.3
sqlparse==0.3.0
Unidecode==1.0.23
